- Added armv6l from https://github.com/Gatux/meteor/commit/877df973012c4a509b6d03fa833a0562e3e9a722   and additionally armv7l and arm64.

+++ scripts/generate-dev-bundle.sh
@@ -67,20 +67,35 @@ fi
 
 if [ $ARCH = "i686" ]; then
   MONGO_VERSION=$MONGO_VERSION_32BIT
+
+
+extractNodeFromTarGz || downloadOfficialNode || downloadNodeFromS3
+
+if [ "$ARCH" == "armv6l" ] ; then
+    # Download Mongo from 4commerce-technologies-ag
+    MONGO_URL="https://bintray.com/4commerce-technologies-ag/meteor-universal/download_file?file_path=arm_dev_bundles%2Fmongo_Linux_armv6l_v2.6.7.tar.gz"
+    echo "Downloading Mongo from ${MONGO_URL}"
+    curl -L "${MONGO_URL}" | tar zx
+else
+    # Download Mongo from mongodb.com. Will download a 64-bit version of Mongo
+    # by default. Will download a 32-bit version of Mongo if using a 32-bit based
+    # OS.
+    MONGO_VERSION=$MONGO_VERSION_64BIT
+    if [ $ARCH = "i686" ]; then
+      MONGO_VERSION=$MONGO_VERSION_32BIT
+    fi
+    MONGO_NAME="mongodb-${OS}-${ARCH}-${MONGO_VERSION}"
+    MONGO_TGZ="${MONGO_NAME}.tgz"
+    MONGO_URL="http://fastdl.mongodb.org/${OS}/${MONGO_TGZ}"
+    echo "Downloading Mongo from ${MONGO_URL}"
+    curl "${MONGO_URL}" | tar zx
+
+    # Put Mongo binaries in the right spot (mongodb/bin)
+    mkdir -p "mongodb/bin"
+    mv "${MONGO_NAME}/bin/mongod" "mongodb/bin"
+    mv "${MONGO_NAME}/bin/mongo" "mongodb/bin"
+    rm -rf "${MONGO_NAME}"
 fi
-
-MONGO_NAME="mongodb-${OS}-${ARCH}-${MONGO_VERSION}"
-MONGO_NAME_SSL="mongodb-${OS}${MONGO_SSL}-${ARCH}-${MONGO_VERSION}"
-MONGO_TGZ="${MONGO_NAME_SSL}.tgz"
-MONGO_URL="http://fastdl.mongodb.org/${OS}/${MONGO_TGZ}"
-echo "Downloading Mongo from ${MONGO_URL}"
-curl "${MONGO_URL}" | tar zx
-
-# Put Mongo binaries in the right spot (mongodb/bin)
-mkdir -p "mongodb/bin"
-mv "${MONGO_NAME}/bin/mongod" "mongodb/bin"
-mv "${MONGO_NAME}/bin/mongo" "mongodb/bin"
-rm -rf "${MONGO_NAME}"
 
 # export path so we use the downloaded node and npm
 export PATH="$DIR/bin:$PATH"
